Move code-actions into sub-packages
