+++ language-server/modules/langserver-core/src/main/java/org/ballerinalang/langserver/codeaction/providers/docs/AddAllDocumentationCodeAction.java
@@ -13,40 +13,39 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.ballerinalang.langserver.codeaction.providers;
+package org.ballerinalang.langserver.codeaction.providers.docs;
 
 import org.ballerinalang.annotation.JavaSPIService;
-import org.ballerinalang.langserver.command.executors.CreateTestExecutor;
+import org.ballerinalang.langserver.codeaction.providers.AbstractCodeActionProvider;
+import org.ballerinalang.langserver.command.executors.AddAllDocumentationExecutor;
 import org.ballerinalang.langserver.common.constants.CommandConstants;
 import org.ballerinalang.langserver.commons.CodeActionContext;
 import org.ballerinalang.langserver.commons.codeaction.CodeActionNodeType;
 import org.ballerinalang.langserver.commons.command.CommandArgument;
 import org.eclipse.lsp4j.CodeAction;
 import org.eclipse.lsp4j.Command;
-import org.eclipse.lsp4j.Position;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
-
 /**
- * Code Action for generating test case for function.
+ * Code Action provider for adding all documentation for top level items.
  *
- * @since 1.2.0
+ * @since 1.1.1
  */
 @JavaSPIService("org.ballerinalang.langserver.commons.codeaction.spi.LSCodeActionProvider")
-public class CreateFunctionTestCodeAction extends AbstractCodeActionProvider {
-    public CreateFunctionTestCodeAction() {
-        super(Collections.singletonList(CodeActionNodeType.FUNCTION));
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public boolean isEnabled() {
-        return false;
+public class AddAllDocumentationCodeAction extends AbstractCodeActionProvider {
+    public AddAllDocumentationCodeAction() {
+        super(Arrays.asList(CodeActionNodeType.FUNCTION,
+                CodeActionNodeType.OBJECT,
+                CodeActionNodeType.CLASS,
+                CodeActionNodeType.SERVICE,
+                CodeActionNodeType.RESOURCE,
+                CodeActionNodeType.RECORD,
+                CodeActionNodeType.OBJECT_FUNCTION,
+                CodeActionNodeType.CLASS_FUNCTION));
     }
 
     /**
@@ -55,16 +54,11 @@ public class CreateFunctionTestCodeAction extends AbstractCodeActionProvider {
     @Override
     public List<CodeAction> getNodeBasedCodeActions(CodeActionContext context) {
         String docUri = context.fileUri();
-        List<CodeAction> actions = new ArrayList<>();
-        List<Object> args = new ArrayList<>();
-        args.add(new CommandArgument(CommandConstants.ARG_KEY_DOC_URI, docUri));
-        Position position = context.cursorPosition();
-        args.add(new CommandArgument(CommandConstants.ARG_KEY_NODE_LINE, "" + position.getLine()));
-        args.add(new CommandArgument(CommandConstants.ARG_KEY_NODE_COLUMN, "" + position.getCharacter()));
+        CommandArgument docUriArg = new CommandArgument(CommandConstants.ARG_KEY_DOC_URI, docUri);
+        List<Object> args = new ArrayList<>(Collections.singletonList(docUriArg));
 
-        CodeAction action = new CodeAction(CommandConstants.CREATE_TEST_FUNC_TITLE);
-        action.setCommand(new Command(CommandConstants.CREATE_TEST_FUNC_TITLE, CreateTestExecutor.COMMAND, args));
-        actions.add(action);
-        return actions;
+        CodeAction action = new CodeAction(CommandConstants.ADD_ALL_DOC_TITLE);
+        action.setCommand(new Command(CommandConstants.ADD_ALL_DOC_TITLE, AddAllDocumentationExecutor.COMMAND, args));
+        return Collections.singletonList(action);
     }
 }
